ExtendedDates

Documentation for ExtendedDates.

ExtendedDates.Semester

ExtendedDates.Semester

ExtendedDates.Semester

ExtendedDates.dayofquarter

ExtendedDates.firstdayofsemester

ExtendedDates.lastdayofsemester

ExtendedDates.semester

ExtendedDates.semesterofyear

ExtendedDates.Semester — Method

Semester(v)

Construct a Semester object with the given v value. Input must be losslessly convertible to an Int64.

ExtendedDates.Semester — Method

Semester(dt::DateTime) -> Semester

The semester part of a DateTime as a Semester.

ExtendedDates.Semester — Method

Semester(dt::Date) -> Semester

The semester part of a Date as a Semester.

ExtendedDates.dayofquarter — Method

dayofquarter(dt::TimeType) -> Int

Return the day of the current quarter of dt. Range of value is 1:92.

ExtendedDates.firstdayofsemester — Method

firstdayofsemester(dt::TimeType) -> TimeType

Adjusts dt to the first day of its semester.

Examples

julia> ExtendedDates.firstdayofsemester(DateTime("1996-05-20"))
1996-01-01T00:00:00

julia> ExtendedDates.firstdayofsemester(DateTime("1996-08-20"))
1996-07-01T00:00:00

ExtendedDates.lastdayofsemester — Method

lastdayofsemester(dt::TimeType) -> TimeType

Adjusts dt to the last day of its semester.

Examples

julia> ExtendedDates.lastdayofsemester(DateTime("1996-05-20"))
1996-06-30T00:00:00

julia> ExtendedDates.lastdayofsemester(DateTime("1996-08-20"))
1996-12-31T00:00:00

ExtendedDates.semester — Method

semester(dt::TimeType) -> Int64

The semester of a Date or DateTime as an Int64.

ExtendedDates.semesterofyear — Method

semesterofyear(dt::TimeType) -> Int

Return the semester that dt resides in. Range of value is 1:4.