ExtendedDates

Documentation for ExtendedDates.

ExtendedDates.Semester

ExtendedDates.Semester

ExtendedDates.Semester

Dates.year

ExtendedDates.dayofsemester

ExtendedDates.epoch

ExtendedDates.firstdayofsemester

ExtendedDates.lastdayofsemester

ExtendedDates.period

ExtendedDates.semester

ExtendedDates.semesterofyear

ExtendedDates.subperiod

ExtendedDates.Semester — Method

Semester(v)

Construct a Semester object with the given v value. Input must be losslessly convertible to an Int64.

ExtendedDates.Semester — Method

Semester(dt::DateTime) -> Semester

The semester part of a DateTime as a Semester.

ExtendedDates.Semester — Method

Semester(dt::Date) -> Semester

The semester part of a Date as a Semester.

Dates.year — Method

year(::UTInstant{<:Period})

The year of a period.

julia> year(period(Day, 1960, 12))
1960

ExtendedDates.dayofsemester — Method

dayofsemester(dt::TimeType) -> Int

Return the day of the current semester of dt. Range of value is 1:184.

ExtendedDates.epoch — Method

epoch(::Type{Period})

The canonical epoch of a given type of period. For most period types the epoch is Saturday, January 1, year 1.

julia> ExtendedDates.epoch(Quarter)
0001-01-01

julia> ExtendedDates.epoch(Week)
0001-01-01

ExtendedDates.firstdayofsemester — Method

firstdayofsemester(dt::TimeType) -> TimeType

Adjusts dt to the first day of its semester.

Examples

julia> ExtendedDates.firstdayofsemester(DateTime("1996-05-20"))
1996-01-01T00:00:00

julia> ExtendedDates.firstdayofsemester(DateTime("1996-08-20"))
1996-07-01T00:00:00

ExtendedDates.lastdayofsemester — Method

lastdayofsemester(dt::TimeType) -> TimeType

Adjusts dt to the last day of its semester.

Examples

julia> ExtendedDates.lastdayofsemester(DateTime("1996-05-20"))
1996-06-30T00:00:00

julia> ExtendedDates.lastdayofsemester(DateTime("1996-08-20"))
1996-12-31T00:00:00

ExtendedDates.period — Function

period(::Type{<:Period}, year::Integer, subperiod::Integer = 1)

Construct a period from a year, subperiod, and frequency.

These periods are represented as an Int64 number of periods since an epoch defined by the ExtendedDates.epoch function. For most period types the epoch is Saturday, January 1, year 0. For week periods, it is Monday, January 3, year 0.

julia> x = period(Semester, 2022, 1)
4043 semesters

julia> Dates.format(x)
"2022-S1"

julia> Date(x)
2022-01-01

julia> Date(period(Week, 0))
0000-01-03

julia> Date(period(Day, 0))
0000-01-01

ExtendedDates.semester — Method

semester(dt::TimeType) -> Int64

The semester of a Date or DateTime as an Int64.

ExtendedDates.semesterofyear — Method

semesterofyear(dt::TimeType) -> Int

Return the semester that dt resides in. Range of value is 1:2.

ExtendedDates.subperiod — Method

year(::UTInstant{<:Period})

The subperiod of a period within a year. Numbering starts at one.

julia> subperiod(period(Day, 1960, 12))
12

julia> Date(period(Day, 1960, 12))
1960-01-12